[move] BufferScopeの定義を移動
